+++ UltralightNet/Callbacks/ULFileSystemGetFileSizeCallback.cs
@@ -5,7 +5,7 @@ namespace UltralightNet
 	[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 	[return: MarshalAs(UnmanagedType.I1)]
 	public delegate bool ULFileSystemGetFileSizeCallback(
-		int fileHandle,
+		nuint fileHandle,
 		out long result
 	);
 }
