+++ UltralightNet/Callbacks/ULFileSystemOpenFileCallback.cs
@@ -3,12 +3,12 @@ using System.Runtime.InteropServices;
 namespace UltralightNet
 {
 	[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-	public unsafe delegate int ULFileSystemOpenFileCallback__PInvoke__(
+	public unsafe delegate nuint ULFileSystemOpenFileCallback__PInvoke__(
 		ULString* path,
 		[MarshalAs(UnmanagedType.I1)]
 		bool openForWriting
 	);
-	public delegate int ULFileSystemOpenFileCallback(
+	public delegate nuint ULFileSystemOpenFileCallback(
 		string path,
 		bool openForWriting
 	);
